requirements.txt
#win+r ->cmd->cd-> pip install -r requirements.txt
matplotlib>=3.2.2
opencv-python>=4.6.0

numpy>=1.24.3
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1

torch>=1.7.0 # install cuda version for better fps performance
torchvision>=0.8.1 # install cuda version for better fps performance

tqdm>=4.64.0
supervision>=0.6.0
psutil>=5.9.5
thop>=0.1.1

PySide6>=6.4.2
pandas>=2.0.1
lap>=0.4

# flask
flask

# Scheduled Tasks
apscheduler

# cyber style
mplcyberpunk

